// src/QSOCollector/WebApp/src/app/api/models/qso.ts
/* tslint:disable */
export interface Qso {
  mY_NAME?: string;
  qsoId?: number;
  address?: string;
  addresS_INTL?: string;
  age?: string;
  a_INDEX?: string;
  anT_AZ?: string;
  anT_EL?: string;
  anT_PATH?: string;
  arrL_SECT?: string;
  awarD_SUBMITTED?: string;
  awarD_GRANTED?: string;
  band?: string;
  banD_RX?: string;
  call?: string;
  check?: string;
  class?: string;
  clubloG_QSO_UPLOAD_DATE?: string;
  clubloG_QSO_UPLOAD_STATUS?: string;
  cnty?: string;
  comment?: string;
  commenT_INTL?: string;
  cont?: string;
  contacteD_OP?: string;
  contesT_ID?: string;
  country?: string;
  countrY_INTL?: string;
  cqz?: string;
  crediT_SUBMITTED?: string;
  crediT_GRANTED?: string;
  darC_DOK?: string;
  distance?: string;
  dxcc?: string;
  email?: string;
  eQ_CALL?: string;
  eqsL_QSLRDATE?: string;
  eqsL_QSLSDATE?: string;
  eqsL_QSL_RCVD?: string;
  eqsL_QSL_SENT?: string;
  fists?: string;
  fistS_CC?: string;
  forcE_INIT?: string;
  freq?: string;
  freQ_RX?: string;
  gridsquare?: string;
  guesT_OP?: string;
  hrdloG_QSO_UPLOAD_DATE?: string;
  hrdloG_QSO_UPLOAD_STATUS?: string;
  iota?: string;
  iotA_ISLAND_ID?: string;
  ituz?: string;
  k_INDEX?: string;
  lat?: string;
  lon?: string;
  lotW_QSLRDATE?: string;
  lotW_QSLSDATE?: string;
  lotW_QSL_RCVD?: string;
  lotW_QSL_SENT?: string;
  maX_BURSTS?: string;
  mode?: string;
  mS_SHOWER?: string;
  mY_ANTENNA?: string;
  mY_ANTENNA_INTL?: string;
  mY_CITY?: string;
  mY_CITY_INTL?: string;
  mY_CNTY?: string;
  mY_COUNTRY?: string;
  mY_COUNTRY_INTL?: string;
  mY_CQ_ZONE?: string;
  mY_DXCC?: string;
  mY_FISTS?: string;
  mY_GRIDSQUARE?: string;
  mY_IOTA?: string;
  mY_IOTA_ISLAND_ID?: string;
  mY_ITU_ZONE?: string;
  mY_LAT?: string;
  mY_LON?: string;
  stationId?: number;
  mY_NAME_INTL?: string;
  mY_POSTAL_CODE?: string;
  mY_POSTAL_CODE_INTL?: string;
  mY_RIG?: string;
  mY_RIG_INTL?: string;
  mY_SIG?: string;
  mY_SIG_INTL?: string;
  mY_SIG_INFO?: string;
  mY_SIG_INFO_INTL?: string;
  mY_SOTA_REF?: string;
  mY_STATE?: string;
  mY_STREET?: string;
  mY_STREET_INTL?: string;
  mY_USACA_COUNTIES?: string;
  mY_VUCC_GRIDS?: string;
  name?: string;
  namE_INTL?: string;
  notes?: string;
  noteS_INTL?: string;
  nR_BURSTS?: string;
  nR_PINGS?: string;
  operator?: string;
  owneR_CALLSIGN?: string;
  pfx?: string;
  precedence?: string;
  proP_MODE?: string;
  publiC_KEY?: string;
  qrzcoM_QSO_UPLOAD_DATE?: string;
  qrzcoM_QSO_UPLOAD_STATUS?: string;
  qslmsg?: string;
  qslmsG_INTL?: string;
  qslrdate?: string;
  qslsdate?: string;
  qsL_RCVD?: string;
  qsL_RCVD_VIA?: string;
  qsL_SENT?: string;
  qsL_SENT_VIA?: string;
  qsL_VIA?: string;
  qsO_COMPLETE?: string;
  qsO_DATE?: string;
  qsO_DATE_OFF?: string;
  qsO_RANDOM?: string;
  qth?: string;
  qtH_INTL?: string;
  region?: string;
  rig?: string;
  riG_INTL?: string;
  rsT_RCVD?: string;
  rsT_SENT?: string;
  rX_PWR?: string;
  saT_MODE?: string;
  saT_NAME?: string;
  sfi?: string;
  sig?: string;
  siG_INTL?: string;
  siG_INFO?: string;
  siG_INFO_INTL?: string;
  silenT_KEY?: string;
  skcc?: string;
  sotA_REF?: string;
  srx?: string;
  srX_STRING?: string;
  state?: string;
  statioN_CALLSIGN?: string;
  stx?: string;
  stX_STRING?: string;
  submode?: string;
  swl?: string;
  teN_TEN?: string;
  timE_OFF?: string;
  timE_ON?: string;
  tX_PWR?: string;
  uksmg?: string;
  usacA_COUNTIES?: string;
  vE_PROV?: string;
  vucC_GRIDS?: string;
  web?: string;
}